:recycle:  extract rules components to individual files + TS
