+++ packages/desktop-client/src/components/ManageRules.js
@@ -1,6 +1,4 @@
 import React, {
-  forwardRef,
-  memo,
   useState,
   useEffect,
   useRef,
@@ -9,491 +7,26 @@ import React, {
 } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { format as formatDate, parseISO } from 'date-fns';
-
 import { pushModal } from 'loot-core/src/client/actions/modals';
 import { initiallyLoadPayees } from 'loot-core/src/client/actions/queries';
-import q from 'loot-core/src/client/query-helpers';
-import { liveQueryContext } from 'loot-core/src/client/query-hooks';
-import { getPayeesById } from 'loot-core/src/client/reducers/queries';
 import { send } from 'loot-core/src/platform/client/fetch';
 import * as undo from 'loot-core/src/platform/client/undo';
-import { getMonthYearFormat } from 'loot-core/src/shared/months';
 import { mapField, friendlyOp } from 'loot-core/src/shared/rules';
-import { getRecurringDescription } from 'loot-core/src/shared/schedules';
-import { integerToCurrency } from 'loot-core/src/shared/util';
+import { describeSchedule } from 'loot-core/src/shared/schedules';
 
-import useSelected, {
-  useSelectedDispatch,
-  useSelectedItems,
-  SelectedProvider,
-} from '../hooks/useSelected';
-import ArrowRight from '../icons/v0/RightArrow2';
+import useSelected, { SelectedProvider } from '../hooks/useSelected';
 import { theme } from '../style';
 
 import Button from './common/Button';
 import ExternalLink from './common/ExternalLink';
-import LinkButton from './common/LinkButton';
 import Search from './common/Search';
 import Stack from './common/Stack';
 import Text from './common/Text';
 import View from './common/View';
-import { SelectCell, Row, Field, Cell, CellButton, TableHeader } from './table';
-
-let valueStyle = {
-  color: theme.pageTextPositive,
-};
-let SchedulesQuery = liveQueryContext(q('schedules').select('*'));
-
-export function Value({
-  value,
-  field,
-  valueIsRaw,
-  inline = false,
-  data: dataProp,
-  describe = x => x.name,
-}) {
-  let { data, dateFormat } = useSelector(state => {
-    let data;
-    if (dataProp) {
-      data = dataProp;
-    } else {
-      switch (field) {
-        case 'payee':
-          data = state.queries.payees;
-          break;
-        case 'category':
-          data = state.queries.categories.list;
-          break;
-        case 'account':
-          data = state.queries.accounts;
-          break;
-        default:
-          data = [];
-      }
-    }
-
-    return {
-      data,
-      dateFormat: state.prefs.local.dateFormat || 'MM/dd/yyyy',
-    };
-  });
-  let [expanded, setExpanded] = useState(false);
-
-  function onExpand(e) {
-    e.preventDefault();
-    setExpanded(true);
-  }
-
-  function formatValue(value) {
-    if (value == null || value === '') {
-      return '(nothing)';
-    } else if (typeof value === 'boolean') {
-      return value ? 'true' : 'false';
-    } else {
-      switch (field) {
-        case 'amount':
-          return integerToCurrency(value);
-        case 'date':
-          if (value) {
-            if (value.frequency) {
-              return getRecurringDescription(value);
-            }
-            return formatDate(parseISO(value), dateFormat);
-          }
-          return null;
-        case 'month':
-          return value
-            ? formatDate(parseISO(value), getMonthYearFormat(dateFormat))
-            : null;
-        case 'year':
-          return value ? formatDate(parseISO(value), 'yyyy') : null;
-        case 'notes':
-        case 'imported_payee':
-          return value;
-        case 'payee':
-        case 'category':
-        case 'account':
-        case 'rule':
-          if (valueIsRaw) {
-            return value;
-          }
-          if (data && data.length) {
-            let item = data.find(item => item.id === value);
-            if (item) {
-              return describe(item);
-            } else {
-              return '(deleted)';
-            }
-          }
-
-          return '…';
-        default:
-          throw new Error(`Unknown field ${field}`);
-      }
-    }
-  }
-
-  if (Array.isArray(value)) {
-    if (value.length === 0) {
-      return <Text style={valueStyle}>(empty)</Text>;
-    } else if (value.length === 1) {
-      return (
-        <Text>
-          [<Text style={valueStyle}>{formatValue(value[0])}</Text>]
-        </Text>
-      );
-    }
-
-    let displayed = value;
-    if (!expanded && value.length > 4) {
-      displayed = value.slice(0, 3);
-    }
-    let numHidden = value.length - displayed.length;
-    return (
-      <Text style={{ color: theme.tableText }}>
-        [
-        {displayed.map((v, i) => {
-          let text = <Text style={valueStyle}>{formatValue(v)}</Text>;
-          let spacing;
-          if (inline) {
-            spacing = i !== 0 ? ' ' : '';
-          } else {
-            spacing = (
-              <>
-                {i === 0 && <br />}
-                &nbsp;&nbsp;
-              </>
-            );
-          }
-
-          return (
-            <Text key={i}>
-              {spacing}
-              {text}
-              {i === value.length - 1 ? '' : ','}
-              {!inline && <br />}
-            </Text>
-          );
-        })}
-        {numHidden > 0 && (
-          <Text style={valueStyle}>
-            &nbsp;&nbsp;
-            <LinkButton onClick={onExpand} style={valueStyle}>
-              {numHidden} more items...
-            </LinkButton>
-            {!inline && <br />}
-          </Text>
-        )}
-        ]
-      </Text>
-    );
-  } else if (value && value.num1 != null && value.num2 != null) {
-    // An "in between" type
-    return (
-      <Text>
-        <Text style={valueStyle}>{formatValue(value.num1)}</Text> and{' '}
-        <Text style={valueStyle}>{formatValue(value.num2)}</Text>
-      </Text>
-    );
-  } else {
-    return <Text style={valueStyle}>{formatValue(value)}</Text>;
-  }
-}
-
-function ConditionExpression({
-  field,
-  op,
-  value,
-  options,
-  prefix,
-  style,
-  inline,
-}) {
-  return (
-    <View
-      style={[
-        {
-          display: 'block',
-          maxWidth: '100%',
-          color: theme.altPillText,
-          backgroundColor: theme.altPillBackground,
-          borderRadius: 4,
-          padding: '3px 5px',
-          whiteSpace: 'nowrap',
-          overflow: 'hidden',
-          textOverflow: 'ellipsis',
-        },
-        style,
-      ]}
-    >
-      {prefix && <Text>{prefix} </Text>}
-      <Text style={valueStyle}>{mapField(field, options)}</Text>{' '}
-      <Text>{friendlyOp(op)}</Text>{' '}
-      <Value value={value} field={field} inline={inline} />
-    </View>
-  );
-}
-
-function describeSchedule(schedule, payee) {
-  if (payee) {
-    return `${payee.name} (${schedule.next_date})`;
-  } else {
-    return `Next: ${schedule.next_date}`;
-  }
-}
-
-function ScheduleValue({ value }) {
-  let payees = useSelector(state => state.queries.payees);
-  let byId = getPayeesById(payees);
-  let { data: schedules } = SchedulesQuery.useQuery();
-
-  return (
-    <Value
-      value={value}
-      field="rule"
-      data={schedules}
-      describe={schedule => describeSchedule(schedule, byId[schedule._payee])}
-    />
-  );
-}
-
-function ActionExpression({ field, op, value, options, style }) {
-  return (
-    <View
-      style={[
-        {
-          display: 'block',
-          maxWidth: '100%',
-          color: theme.altPillText,
-          backgroundColor: theme.altPillBackground,
-          borderRadius: 4,
-          padding: '3px 5px',
-          whiteSpace: 'nowrap',
-          overflow: 'hidden',
-          textOverflow: 'ellipsis',
-        },
-        style,
-      ]}
-    >
-      {op === 'set' ? (
-        <>
-          <Text>{friendlyOp(op)}</Text>{' '}
-          <Text style={valueStyle}>{mapField(field, options)}</Text>{' '}
-          <Text>to </Text>
-          <Value value={value} field={field} />
-        </>
-      ) : op === 'link-schedule' ? (
-        <>
-          <Text>{friendlyOp(op)}</Text> <ScheduleValue value={value} />
-        </>
-      ) : null}
-    </View>
-  );
-}
-
-let Rule = memo(({ rule, hovered, selected, onHover, onEditRule }) => {
-  let dispatchSelected = useSelectedDispatch();
-  let borderColor = selected ? theme.tableBorderSelected : 'none';
-  let backgroundFocus = hovered;
-
-  return (
-    <Row
-      height="auto"
-      style={{
-        fontSize: 13,
-        zIndex: selected ? 101 : 'auto',
-        borderColor,
-        backgroundColor: selected
-          ? theme.tableRowBackgroundHighlight
-          : backgroundFocus
-          ? theme.tableRowBackgroundHover
-          : theme.tableBackground,
-      }}
-      collapsed="true"
-      onMouseEnter={() => onHover && onHover(rule.id)}
-      onMouseLeave={() => onHover && onHover(null)}
-    >
-      <SelectCell
-        exposed={hovered || selected}
-        focused={true}
-        onSelect={e => {
-          dispatchSelected({ type: 'select', id: rule.id, event: e });
-        }}
-        selected={selected}
-      />
-
-      <Cell name="stage" width={50} plain style={{ color: theme.tableText }}>
-        {rule.stage && (
-          <View
-            style={{
-              alignSelf: 'flex-start',
-              margin: 5,
-              backgroundColor: theme.pillBackgroundSelected,
-              color: theme.pillTextSelected,
-              borderRadius: 4,
-              padding: '3px 5px',
-            }}
-          >
-            {rule.stage}
-          </View>
-        )}
-      </Cell>
-
-      <Field width="flex" style={{ padding: '15px 0' }} truncate={false}>
-        <Stack direction="row" align="center">
-          <View
-            style={{ flex: 1, alignItems: 'flex-start' }}
-            data-testid="conditions"
-          >
-            {rule.conditions.map((cond, i) => (
-              <ConditionExpression
-                key={i}
-                field={cond.field}
-                op={cond.op}
-                inline={true}
-                value={cond.value}
-                options={cond.options}
-                prefix={i > 0 ? friendlyOp(rule.conditionsOp) : null}
-                style={i !== 0 && { marginTop: 3 }}
-              />
-            ))}
-          </View>
-
-          <Text>
-            <ArrowRight
-              style={{ width: 12, height: 12, color: theme.tableText }}
-            />
-          </Text>
-
-          <View
-            style={{ flex: 1, alignItems: 'flex-start' }}
-            data-testid="actions"
-          >
-            {rule.actions.map((action, i) => (
-              <ActionExpression
-                key={i}
-                field={action.field}
-                op={action.op}
-                value={action.value}
-                options={action.options}
-                style={i !== 0 && { marginTop: 3 }}
-              />
-            ))}
-          </View>
-        </Stack>
-      </Field>
-
-      <Cell name="edit" plain style={{ padding: '0 15px', paddingLeft: 5 }}>
-        <Button as={CellButton} onSelect={() => onEditRule(rule)}>
-          Edit
-        </Button>
-      </Cell>
-    </Row>
-  );
-});
-
-let SimpleTable = forwardRef(
-  ({ data, loadMore, style, onHoverLeave, children, ...props }, ref) => {
-    let contentRef = useRef();
-    let contentHeight = useRef();
-    let scrollRef = useRef();
-
-    function onScroll(e) {
-      if (contentHeight.current != null) {
-        if (loadMore && e.target.scrollTop > contentHeight.current - 750) {
-          loadMore();
-        }
-      }
-    }
-
-    useEffect(() => {
-      if (contentRef.current) {
-        contentHeight.current =
-          contentRef.current.getBoundingClientRect().height;
-      } else {
-        contentHeight.current = null;
-      }
-    }, [contentRef.current, data]);
-
-    return (
-      <View
-        style={[
-          {
-            flex: 1,
-            outline: 'none',
-            '& .animated .animated-row': { transition: '.25s transform' },
-          },
-          style,
-        ]}
-        tabIndex="1"
-        data-testid="table"
-      >
-        <View
-          innerRef={scrollRef}
-          style={{ maxWidth: '100%', overflow: 'auto' }}
-          onScroll={onScroll}
-        >
-          <div ref={contentRef} onMouseLeave={onHoverLeave}>
-            {children}
-          </div>
-        </View>
-      </View>
-    );
-  },
-);
-
-function RulesHeader() {
-  let selectedItems = useSelectedItems();
-  let dispatchSelected = useSelectedDispatch();
-
-  return (
-    <TableHeader version="v2" style={{}}>
-      <SelectCell
-        exposed={true}
-        focused={false}
-        selected={selectedItems.size > 0}
-        onSelect={e => dispatchSelected({ type: 'select-all', event: e })}
-      />
-      <Cell value="Stage" width={50} />
-      <Cell value="Rule" width="flex" />
-    </TableHeader>
-  );
-}
-
-function RulesList({
-  rules,
-  selectedItems,
-  hoveredRule,
-  collapsed: borderCollapsed,
-  onHover,
-  onCollapse,
-  onEditRule,
-}) {
-  if (rules.length === 0) {
-    return null;
-  }
-
-  return (
-    <View>
-      {rules.map(rule => {
-        let hovered = hoveredRule === rule.id;
-        let selected = selectedItems.has(rule.id);
-
-        return (
-          <Rule
-            key={rule.id}
-            rule={rule}
-            hovered={hovered}
-            selected={selected}
-            onHover={onHover}
-            onEditRule={onEditRule}
-          />
-        );
-      })}
-    </View>
-  );
-}
+import RulesHeader from './rules/RulesHeader';
+import RulesList from './rules/RulesList';
+import { SchedulesQuery } from './rules/SchedulesQuery';
+import SimpleTable from './rules/SimpleTable';
 
 function mapValue(field, value, { payees, categories, accounts }) {
   if (!value) return '';
