+++ packages/desktop-client/src/components/common/Button.tsx
@@ -1,4 +1,4 @@
-import React, { forwardRef } from 'react';
+import React, { forwardRef, type ElementType } from 'react';
 
 import { css, type CSSProperties } from 'glamor';
 
@@ -19,7 +19,7 @@ type ButtonProps = HTMLPropsWithStyle<HTMLButtonElement> & {
   activeStyle?: CSSProperties;
   textStyle?: CSSProperties;
   bounce?: boolean;
-  as?: 'button';
+  as?: ElementType;
 };
 
 type ButtonType = 'normal' | 'primary' | 'bare';
