+++ packages/desktop-client/src/components/rules/Value.tsx
@@ -1,0 +1,185 @@
+import React, { useState } from 'react';
+import { useSelector } from 'react-redux';
+
+import { format as formatDate, parseISO } from 'date-fns';
+
+import { getMonthYearFormat } from 'loot-core/src/shared/months';
+import { getRecurringDescription } from 'loot-core/src/shared/schedules';
+import { integerToCurrency } from 'loot-core/src/shared/util';
+
+import { theme } from '../../style';
+import LinkButton from '../common/LinkButton';
+import Text from '../common/Text';
+
+let valueStyle = {
+  color: theme.pageTextPositive,
+};
+
+type ValueProps<T> = {
+  value: T;
+  field: unknown;
+  valueIsRaw?: boolean;
+  inline?: boolean;
+  data?: unknown;
+  describe?: (item: T) => string;
+};
+
+export default function Value<T>({
+  value,
+  field,
+  valueIsRaw,
+  inline = false,
+  data: dataProp,
+  // @ts-expect-error fix this later
+  describe = x => x.name,
+}: ValueProps<T>) {
+  let { data, dateFormat } = useSelector(state => {
+    let data;
+    if (dataProp) {
+      data = dataProp;
+    } else {
+      switch (field) {
+        case 'payee':
+          data = state.queries.payees;
+          break;
+        case 'category':
+          data = state.queries.categories.list;
+          break;
+        case 'account':
+          data = state.queries.accounts;
+          break;
+        default:
+          data = [];
+      }
+    }
+
+    return {
+      data,
+      dateFormat: state.prefs.local.dateFormat || 'MM/dd/yyyy',
+    };
+  });
+  let [expanded, setExpanded] = useState(false);
+
+  function onExpand(e) {
+    e.preventDefault();
+    setExpanded(true);
+  }
+
+  function formatValue(value) {
+    if (value == null || value === '') {
+      return '(nothing)';
+    } else if (typeof value === 'boolean') {
+      return value ? 'true' : 'false';
+    } else {
+      switch (field) {
+        case 'amount':
+          return integerToCurrency(value);
+        case 'date':
+          if (value) {
+            if (value.frequency) {
+              return getRecurringDescription(value);
+            }
+            return formatDate(parseISO(value), dateFormat);
+          }
+          return null;
+        case 'month':
+          return value
+            ? formatDate(parseISO(value), getMonthYearFormat(dateFormat))
+            : null;
+        case 'year':
+          return value ? formatDate(parseISO(value), 'yyyy') : null;
+        case 'notes':
+        case 'imported_payee':
+          return value;
+        case 'payee':
+        case 'category':
+        case 'account':
+        case 'rule':
+          if (valueIsRaw) {
+            return value;
+          }
+          if (data && data.length) {
+            let item = data.find(item => item.id === value);
+            if (item) {
+              return describe(item);
+            } else {
+              return '(deleted)';
+            }
+          }
+
+          return '…';
+        default:
+          throw new Error(`Unknown field ${field}`);
+      }
+    }
+  }
+
+  if (Array.isArray(value)) {
+    if (value.length === 0) {
+      return <Text style={valueStyle}>(empty)</Text>;
+    } else if (value.length === 1) {
+      return (
+        <Text>
+          [<Text style={valueStyle}>{formatValue(value[0])}</Text>]
+        </Text>
+      );
+    }
+
+    let displayed: unknown[] = value;
+    if (!expanded && value.length > 4) {
+      displayed = value.slice(0, 3);
+    }
+    let numHidden = value.length - displayed.length;
+    return (
+      <Text style={{ color: theme.tableText }}>
+        [
+        {displayed.map((v, i) => {
+          let text = <Text style={valueStyle}>{formatValue(v)}</Text>;
+          let spacing;
+          if (inline) {
+            spacing = i !== 0 ? ' ' : '';
+          } else {
+            spacing = (
+              <>
+                {i === 0 && <br />}
+                &nbsp;&nbsp;
+              </>
+            );
+          }
+
+          return (
+            <Text key={i}>
+              {spacing}
+              {text}
+              {i === value.length - 1 ? '' : ','}
+              {!inline && <br />}
+            </Text>
+          );
+        })}
+        {numHidden > 0 && (
+          <Text style={valueStyle}>
+            &nbsp;&nbsp;
+            <LinkButton onClick={onExpand} style={valueStyle}>
+              {numHidden} more items...
+            </LinkButton>
+            {!inline && <br />}
+          </Text>
+        )}
+        ]
+      </Text>
+    );
+    // @ts-expect-error Fix typechecker here
+  } else if (value && value.num1 != null && value.num2 != null) {
+    // An "in between" type
+    // @ts-expect-error Fix typechecker here
+    const { num1, num2 } = value;
+    return (
+      <Text>
+        <Text style={valueStyle}>{formatValue(num1)}</Text> and{' '}
+        <Text style={valueStyle}>{formatValue(num2)}</Text>
+      </Text>
+    );
+  } else {
+    return <Text style={valueStyle}>{formatValue(value)}</Text>;
+  }
+}
