+++ packages/loot-core/src/types/models/index.d.ts
@@ -2,5 +2,6 @@ export type * from './account';
 export type * from './category';
 export type * from './category-group';
 export type * from './payee';
+export type * from './rule';
 export type * from './schedule';
 export type * from './transaction';
