+++ packages/loot-core/src/types/models/rule.d.ts
@@ -2,7 +2,22 @@ export interface RuleEntity {
   id: string;
   stage: string;
   conditions_op: string;
-  conditions: unknown;
-  actions: unknown;
+  conditions: RuleConditionEntity[];
+  actions: RuleActionEntity[];
   tombstone: boolean;
+}
+
+interface RuleConditionEntity {
+  field: unknown;
+  op: unknown;
+  value: unknown;
+  options: unknown;
+  conditionsOp: unknown;
+}
+
+interface RuleActionEntity {
+  field: unknown;
+  op: unknown;
+  value: unknown;
+  options: unknown;
 }
